requirements.txt
fastapi
uvicorn[standard]
asyncpg
watchdog
pydantic
python-dotenv
geoip2
